@@ -12,10 +12,10 @@
     }
   },
   "dependencies": {
-    "@reactivex/rxjs": "^5.0.0-rc.4",
+    "@reactivex/rxjs": "^5.0.2",
     "browserify": "^13.1.1",
     "electron-sudo": "^4.0.12",
-    "electron-window-state": "^4.0.0",
+    "electron-window-state": "^4.0.1",
     "eventemitter2": "^2.2.1",
     "http-mitm-proxy": "^0.5.1",
     "ifconfig": "git+https://github.com/femto113/node-ifconfig.git",
@@ -23,31 +23,31 @@
     "minimatch": "^3.0.3",
     "react": "^15.4.1",
     "react-dom": "^15.4.1",
-    "react-redux": "^4.4.6",
+    "react-redux": "^5.0.1",
     "redux": "^3.6.0",
-    "regenerator-runtime": "^0.10.0",
+    "regenerator-runtime": "^0.10.1",
     "remote": "^0.2.6",
     "typescript-dddbase": "0.0.2"
   },
   "devDependencies": {
-    "@types/electron": "^1.4.27",
-    "@types/eventemitter2": "^2.2.0",
+    "@types/electron": "^1.4.30",
+    "@types/eventemitter2": "^2.2.1",
     "@types/mime": "0.0.29",
     "@types/minimatch": "^2.0.29",
-    "@types/mocha": "^2.2.33",
-    "@types/node": "^6.0.51",
-    "@types/react": "^0.14.50",
+    "@types/mocha": "^2.2.34",
+    "@types/node": "^6.0.53",
+    "@types/react": "^0.14.55",
     "@types/react-dom": "^0.14.19",
-    "@types/react-redux": "^4.4.33",
+    "@types/react-redux": "^4.4.35",
     "@types/redux": "^3.6.31",
-    "electron": "^1.4.8",
-    "electron-builder": "^8.6.0",
+    "electron": "^1.4.13",
+    "electron-builder": "^10.12.0",
     "fixpack": "^2.3.1",
     "mocha": "^3.2.0",
     "mock-fs": "^3.12.1",
     "mock-require": "^2.0.0",
-    "npm-check-updates": "^2.8.7",
-    "typescript": "^2.0.10"
+    "npm-check-updates": "^2.8.9",
+    "typescript": "^2.1.4"
   },
   "homepage": "https://github.com/kyo-ago/lifter/",
   "keywords": [
