@@ -1,15 +1,14 @@
-import {Url} from "url";
+import * as Url from "url";
 import {BaseValueObject} from "../../base/value-object";
-const NodeUrl = require('url');
 
 export class ClientRequestUrl extends BaseValueObject<string> {
-    private reqestUrl: Url;
+    private reqestUrl: Url.Url;
 
     constructor(
         _value: string,
     ) {
         super(_value);
-        this.reqestUrl = NodeUrl.parse(_value);
+        this.reqestUrl = Url.parse(_value);
     }
 
     getPathname() {
